requirements.txt
jumanji
dm-haiku
hydra-core==1.3
neptune-client==0.16.15
optax>=0.1.4
rlax>=0.1.4
tensorboardX==2.5.1
tqdm>=4.64.1
minizinc
lark-parser
tensorboard
pillow==9.0.0
jax[cuda12]